requirements.txt
fastapi
uvicorn
async-exit-stack 
async-generator
textblob
mysql-connector
wheel
pandas
plotly_express
nltk
tweepy